refactor: keep the core project more aot-compatible
