+++ src/WeihanLi.Core/Helpers/EnumHelper.cs
@@ -1,8 +1,6 @@
 ﻿using System.ComponentModel;
-using System.Diagnostics.CodeAnalysis;
 using System.Reflection;
 using WeihanLi.Common.Models;
-using WeihanLi.Extensions;
 
 namespace WeihanLi.Common.Helpers;
 
@@ -17,18 +15,7 @@ public static class EnumHelper
             Id = Convert.ToInt32(Enum.Parse(enumType, name))
         });
     }
-
-    [RequiresUnreferencedCode("Generic TypeConverters may require the generic types to be annotated. For example, NullableConverter requires the underlying type to be DynamicallyAccessedMembers All.")]
-    public static IReadOnlyList<IdNameModel<TValue>> ToIdNameList<[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicFields)] TEnum, TValue>() where TEnum : Enum
-    {
-        var enumType = typeof(TEnum);
-        return Array.ConvertAll(Enum.GetNames(enumType), name => new IdNameModel<TValue>()
-        {
-            Id = Enum.Parse(enumType, name).To<TValue>(),
-            Name = name,
-        });
-    }
-
+    
     public static IReadOnlyList<IdNameDescModel> ToIdNameDescList<[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicFields)] TEnum>() where TEnum : Enum
     {
         var enumType = typeof(TEnum);
@@ -36,18 +23,6 @@ public static class EnumHelper
         {
             Name = name,
             Id = Convert.ToInt32(Enum.Parse(enumType, name)),
-            Description = GetDescription(enumType.GetField(name))
-        });
-    }
-
-    [RequiresUnreferencedCode("Generic TypeConverters may require the generic types to be annotated. For example, NullableConverter requires the underlying type to be DynamicallyAccessedMembers All.")]
-    public static IReadOnlyList<IdNameDescModel<TValue>> ToIdNameDescList<[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicFields)] TEnum, TValue>() where TEnum : Enum
-    {
-        var enumType = typeof(TEnum);
-        return Array.ConvertAll(Enum.GetNames(enumType), converter: name => new IdNameDescModel<TValue>()
-        {
-            Id = Enum.Parse(enumType, name).To<TValue>(),
-            Name = name,
             Description = GetDescription(enumType.GetField(name))
         });
     }
